+++ vue/src/components/pages/pre_start_submissions/PreStartSubmissionsPage.vue
@@ -34,6 +34,7 @@
         :assets="assets"
         :operators="operators"
         :icons="icons"
+        @refresh="onRefresh()"
       />
     </hxCard>
   </div>