+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fixed autosize textarea breaking words
 - fixed drag and drag being presented as a table
 - Added multiple assets for one device notification
+- fixed pre-start section/control ordering
 
 ## 0.10.0
 - added centralised timezone selector